+++ graphql/app.py
@@ -2,7 +2,7 @@ from flask import Flask
 from flask_graphql import GraphQLView
 
 from models import db_session
-from schema import schema, Hierarchy, Comment, Expert
+from schema import schema, Hierarchy, Comment, Expert, Longname, NodcIds
 
 app = Flask(__name__)
 app.debug = True