+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Initial models and schema for the ITIS database.
+- Added models and schemas for longnames and nodc_ids.